handle conversions to int32 and int64 with SGObject::put

+++ src/shogun/mathematics/Seedable.h
@@ -21,14 +21,14 @@ namespace shogun
 		    std::enable_if_t<std::is_base_of<
 		        SGObject, typename std::remove_pointer<T>::type>::value>* =
 		        nullptr>
-		static inline void seed(T* object, int32_t seed)
+		static inline void seed(T* object, int64_t seed)
 		{
 			if (object->has(kSeed))
 				object->put(kSeed, seed);
 		}
 	} // namespace random
 
-	static inline void seed_callback(SGObject*, int32_t);
+	static inline void seed_callback(SGObject*, int64_t);
 
 	template <typename Parent>
 	class Seedable : public Parent
@@ -65,7 +65,7 @@ namespace shogun
 		int64_t m_seed;
 	};
 
-	static inline void seed_callback(SGObject* obj, int32_t seed)
+	static inline void seed_callback(SGObject* obj, int64_t seed)
 	{
 		obj->for_each_param_of_type<SGObject*>(
 		    [&](const std::string& name, SGObject** param) {
